pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "uQCme"
version = "0.1.0"
description = "Microbial Quality Control Dashboard"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=1.5.0",
    "plotly>=5.15.0",
    "PyYAML>=6.0",
    "openpyxl>=3.0.0",
]

[project.scripts]
uqcme = "uQCme.cli:main"
uqcme-dashboard = "uQCme.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
uQCme = ["defaults/*"]
